backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
beanie==1.23.6
pymongo==4.6.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
pandas>=2.2.0
numpy>=1.26.2
scikit-learn>=1.4.0
scikit-optimize>=0.9.0
xgboost==2.0.2
shap==0.43.0
lime==0.2.0.1
scipy==1.11.4
cryptography==41.0.7
pycryptodome==3.19.0
ipfshttpclient==0.8.0a2
redis==5.0.1
celery==5.3.4
reportlab==4.0.7
pypdf2==3.0.1
aiofiles==23.2.1
httpx==0.25.2
requests>=2.31.0
pytz==2024.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytrials==0.3.0
email-validator==2.3.0

